butterfly and sisl renaming
