+++ pettingzoo/sisl/pursuit_v4.py
@@ -1,1 +1,1 @@
-from .pursuit.pursuit import ManualPolicy, env, parallel_env, raw_env
+from .pursuit.pursuit_env import ManualPolicy, env, parallel_env, raw_env
