+++ pettingzoo/sisl/waterworld_v3.py
@@ -1,1 +1,1 @@
-from .waterworld.waterworld import env, parallel_env, raw_env
+from .waterworld.waterworld_env import env, parallel_env, raw_env
